+++ src/main.test.ts
@@ -1,0 +1,14 @@
+// import Magami from './main'
+
+
+// import { describe, test, it, expect } from "vitest";
+
+// const magami = new Magami()
+
+// describe("main", () => {
+//     test('init()', () => {
+//         it('should receive two string parameter', () => {
+//             expect(init('')).toBe('string')
+//         })
+//     })
+// })